Build 33
- Trinket parsing is now done as the request is built.  The parsed
trinkets are stored in a new attribute, parsed_trinket_dimension.
- Added class & spec to the table in mypulls.html
- Fixed a bug where editing an existing request would start with an
empty trinket dimension.
- Fixed a bug that prevented trinket dimensions from being removed.
- Each spec in a request now generates a separate pull.
- Updated task worker to use refactored requests.py
- Taskqueue now permanently fails after 10 attempts.
- Completed refactor of requests.py
- Cleaned up visual presentation of mypulls.html
