+++ queue.yaml
@@ -1,0 +1,5 @@
+queue:
+- name: default
+  rate: 5/s
+  retry_parameters:
+    task_retry_limit: 10
