+++ requests.py
@@ -2,300 +2,37 @@
 
 #Module to handle structuring and executing requests against external apis.
 
-#Here's the deal: yes, my use of "parameter" and "dimension" is inconsistent as
-#hell.  I totally plan to fix it after I have a working prototype for people to
-#break.
-
 import json
 import os
 import logging
 import math
 
 import main
+import cloudstorage as gcs
 
 from google.appengine.ext import ndb
 from google.appengine.api import urlfetch
 
+DEFAULT_LIMIT = 5000
+    
+class Selection(ndb.Model):
+    dimension = ndb.KeyProperty()
+    parameter = ndb.KeyProperty()
 
-class PullTimeoutError(Exception):
+class Filter(ndb.Model):
+    name = ndb.StringProperty()
+    string = ndb.StringProperty()
+    selections = ndb.StructuredProperty(Selection, repeated=True)
+    failed_pages = ndb.IntegerProperty(repeated=True)
+
+class PullFailedError(Exception):
     def __init__(self, msg):
         self.msg = msg
-    
-    
-def rankings_pull_filtered(pull):
-    #This is the function that should be the core of any request pull.  It uses
-    #the other functions below to actually structure, scale, and implement the
-    #pull.
-    
-    result = []
-    request = pull.request.get()
-    encounter_id = pull.encounter,
-    pull_parameters = {
-        "metric": pull.metric,
-        "difficulty": pull.difficulty,
-        "class": request.character_class,
-        }
-    
-    for spec in request.specialization:
-        dimensions = {}
-        pull_parameters['spec'] = spec
-        
-        #Pull NDB objects from stored keys and store them in a dictionary.
-        ''' Dictionary goes out as:
-            {
-                "dimension name": {
-                    "parameter name": {
-                        "include": [spell_id, spell_id] OR None,
-                        "exclude": [spell_id, spell_id] OR None
-                        }
-                    }
-                }
-            '''
-        dimensions_objects = ndb.get_multi(request.dimensions)
-        for dimension in dimensions_objects:
-            dimensions[dimension.name] = {}
-            dim_parameters = ndb.get_multi(dimension.parameters)
-            for parameter in dim_parameters:
-                dimensions[dimension.name][parameter.name] = {
-                    "include": parameter.include,
-                    "exclude": parameter.exclude
-                    }
-        #Add a dimension for trinkets if the request analyzes trinkets.
-        if request.trinket_dimension != None:
-            trinkets = main.Dimension.get_by_id(request.trinket_dimension.id())
-            dimensions["Trinkets"] = build_trinket_dimensions(trinkets)
-        
-        #Build a list of filters to pull against from WCL, then send the pulls.
-        filters = build_filters(dimensions)
-        main.vislog(filters)
-        for filter in filters:
-            get_filter = []
-            pull_parameters['filter'] = filter['filter']
-            
-            logging.info("Pulling ranks for %s" % filter["name"])
-            
-            try:
-                get_filter = rankings_pull_all_pages(pull, pull_parameters)
-            except PullTimeoutError as e:
-                logging.error(e.msg)
-                return None
-            
-            for item in get_filter:
-                for dimension in filter["dimensions"]:
-                    item[dimension] = filter["dimensions"][dimension]
-            result += get_filter
-    
-    return result
-    
-    
-def build_filters(dimensions):
-    #Take a dictionary of dimensions and create a set of unique filters for
-    #each permution of combining parameters from those dimensions.
-    filter_count = 1
-    filter_options = []
-    filter_counter = []
-    filter_reference = {}
-    dimension_reference = {}
-    result = []
-    dimension_counter = 0
-    for dimension in dimensions:
-        filter_count *= len(dimensions[dimension])
-        filter_options.append(len(dimensions[dimension]))
-        filter_counter.append(0)
-        item_counter = 0
-        current = len(filter_counter) - 1
-        dimension_reference[dimension_counter] = dimension
-        filter_reference[current] = {}
-        for item in dimensions[dimension]:
-            filter_reference[current][item_counter] = item
-            item_counter += 1
-        dimension_counter += 1
-            
-    #filter dictionary format-
-    #name:
-    #dimensions: [dimension1, dimension2...]
-    #filter: 
-    
-    for i in range(filter_count):
-        for j in range(len(dimensions)):
-            if filter_counter[j] < filter_options[j]:
-                include = "abilities"
-                exclude = "noabilities"
-                filter_name = ""
-                filter_built = {"dimensions":{}}
-                
-                for k in range(len(dimensions)):
-                    filter_k = filter_reference[k][filter_counter[k]]
-                    
-                    filter_name += filter_k
-                    dimension_name = dimension_reference[k]
-                    filter_built["dimensions"][dimension_name] = filter_k
-                    if k < len(dimensions) - 1:
-                        filter_name += "|"
-                        
-                    current_dimension = dimensions[dimension_name][filter_k]
-                    if current_dimension["include"] != None:
-                        for spellID in current_dimension["include"]:
-                            include += "." + str(spellID)
-                    if current_dimension["exclude"] != None:
-                        for spellID in current_dimension["exclude"]:
-                            exclude += "." + str(spellID)
-                    
-                if len(include) == 9:
-                    include = ""
-                else:
-                    if len(exclude) == 11:
-                        exclude = ""
-                    else:
-                        exclude = "|" + exclude
-                        
-                filter_built["filter"] = include + exclude
-                filter_counter[0] += 1
-                filter_built["name"] = filter_name
-                result.append(filter_built)
-                break
-            else:
-                filter_counter[j] = 0
-                filter_counter[j + 1] += 1
-    
-    return result
-        
-    
-def build_trinket_dimensions(trinkets_dimension):
-    #Take a set of trinket options and generate a dimensions dictionary to be
-    #appended to the other dimensions.
-    ''' Dictionary goes out as:
-            {
-                "Trinket1|Trinket2": {
-                    "include": [1, 2, ... n],
-                    "exclude": [1, 2, ... n]
-                    }
-                }
-            '''
-    slot = [0,1]
-    result = {}
-    trinkets_exclude_all = []
-    
-    #We need to create a list of all trinket effects we're considering.  During
-    #the loop, we'll remove the other trinket's effect from this list for each
-    #trinket considered.
-    trinkets = ndb.get_multi(trinkets_dimension.parameters)
-    for trinket in trinkets:
-        if trinket.include != None:
-            trinkets_exclude_all += trinket.include
-    if trinkets_dimension.other_trinkets == True:
-        result["Both Other Trinkets"] = {}
-        result["Both Other Trinkets"]["include"] = None
-        result["Both Other Trinkets"]["exclude"] = trinkets_exclude_all
-        trinkets.append(main.Parameter(name="Other Trinkets"))
-    
-    #We use a theorem one stars and bars formula to determine the number of
-    #possible combinations of trinkets.  Because k is always 2 (number of 
-    #trinket slots,) the formula can be reduced significantly.        
-    dimensions = (len(trinkets)**2 - len(trinkets)) / 2
-    
-    for dimension in range(dimensions):
-    
-        #If one of the trinkets is "Other Trinkets", apply an "exclude" to it
-        #equal to the "include" of every trinket except the paired counterpart.
-        if trinkets[slot[0]].name == "Other Trinkets":
-            other_trinkets_exclude = list(trinkets_exclude_all)
-            for i in range(len(trinkets[slot[1]].include)):
-                other_trinkets_exclude.remove(trinkets[slot[1]].include[i])
-            trinkets[slot[0]].exclude = other_trinkets_exclude
-        elif trinkets[slot[1]].name == "Other Trinkets":
-            other_trinkets_exclude = list(trinkets_exclude_all)
-            for i in range(len(trinkets[slot[0]].include)):
-                other_trinkets_exclude.remove(trinkets[slot[0]].include[i])
-            trinkets[slot[1]].exclude = other_trinkets_exclude
         
         
-        pair = {"include": [], "exclude": []}
-        name = trinkets[slot[0]].name + "|" + trinkets[slot[1]].name
-        if trinkets[slot[0]].include != None:
-            pair["include"] += trinkets[slot[0]].include
-        if trinkets[slot[1]].include != None:
-            pair["include"] += trinkets[slot[1]].include
-        if trinkets[slot[0]].exclude != None:
-            pair["exclude"] += trinkets[slot[0]].exclude
-        if trinkets[slot[1]].exclude != None:
-            pair["exclude"] += trinkets[slot[1]].exclude
-        
-        if len(pair["include"]) == 0:
-            pair["include"] = None
-        if len(pair["exclude"]) == 0:
-            pair["exclude"] = None
-        
-        result[name] = pair
-        
-        #increment trinket pairings
-        if slot[1] + 1 < len(trinkets):
-            slot[1] += 1
-        elif slot[0] + 1 < slot[1]:
-            slot[0] += 1
-            slot[1] = slot[0] + 1
-            
-    return result
-        
-    
-def rankings_pull_all_pages(pull, parameters):
-    #Take a set of dimensions(w/ a unique set of parameters,) and pull all
-    #pages of applicable rankings.
-    
-    limit = 5000
-    parameters['limit'] = limit
-    
-    for attempt in range(50):
-        logging.info("Pull %d on first page" % (attempt + 1))
-        first_pull = rankings_pull(pull.encounter, parameters)
-        if first_pull != None:
-            break
-    else:
-        logging.info("Could not retreive initial page.")
-        raise PullTimeoutError("Inital pull failed after 50 attempts.")
-        
-    logging.info("Total rankings found: %d" % first_pull["total"])
-    total_pages = int(math.ceil(first_pull["total"] / limit)) + 1
-    result = first_pull["rankings"]
-    
-    for i in range(2, total_pages + 1):
-        page = i
-        for j in range(50):
-            logging.info("Pull %d on page %d of %d" % ((j + 1), i, total_pages))
-            next_pull = rankings_pull(pull.encounter, parameters)
-            if next_pull != None:
-                result += next_pull["rankings"]
-                break
-        else:
-            raise PullTimeoutError("Pull %d of %d failed after 50 attempts." %\
-                                   (i, total_pages))
-                
-    return result
-    
-    
-def rankings_pull(encounter_id, parameters):
-    #Pull a single page of ranks.
-    key = json_pull("apikeys.json")["WCL"]["key"]
-    url_base = "https://www.warcraftlogs.com:443/v1/rankings/encounter/"
-    url_middle = str(encounter_id) + "?"
-    
-    #In this loop, we take each parameter and stick it into the request url.
-    for parameter in parameters:
-        if parameters[parameter] != None:
-            url_middle += "&" + parameter + "=" + str(parameters[parameter])
-    
-    url = url_base + url_middle + "&api_key=" + key
-    main.vislog(url)
-        
-    try:
-        urlfetch.set_default_fetch_deadline(60)
-        response = urlfetch.fetch(url)
-        result = json.loads(response.content)
-    except urlfetch.Error:
-        logging.error("Pull failed.")
-        return None
-    
-    return result
+class ProcessFailureError(Exception):
+    def __init__(self, msg):
+        self.msg = msg
     
                       
 def json_pull(dct):
@@ -317,6 +54,263 @@ def static_request(site, type):
         return None
         
     return result
+
+
+# Work a pull:
+def work_pull(pull):
+    request = pull.request.get()
+    dimensions = ndb.get_multi(request.dimensions)
+    # If trinkets are being considered:
+    if request.trinket_dimension != None:
+        trinket_dimension = request.parsed_trinket_dimension.get()
+        dimensions.append(trinket_dimension)
+    failed_pulls = []
+    results = []
+    # Flag the pull as processing.
+    pull.status = 'Processing'
+    # Build a base query string using the shared (non-filter) arguments *except* spec.
+    query_base = ("https://www.warcraftlogs.com:443/v1/rankings/encounter/" +
+                  str(pull.encounter) + "?metric=" + pull.metric +
+                  "&difficulty=" + str(pull.difficulty) +
+                  "&class=" + str(request.character_class) +
+                  "&limit=" + str(DEFAULT_LIMIT) +
+                  "&spec=" + str(pull.spec) +
+                  "&api_key=" + json_pull("apikeys.json")["WCL"]["key"])
+    
+    # Build a list of "unpacked" pseudo dimensions - dictionaries in the format
+    '''{
+        'name': dimension.name,
+        'key': dimension.key,
+        'parameters': [Parameter, Parameter, ...],
+        }'''
+    dimensions_pseudo = []
+    for dimension in dimensions:
+        new_pseudo = {'name':dimension.name, 
+                      'key': dimension.key,
+                      'parameters':ndb.get_multi(dimension.parameters)}
+        dimensions_pseudo.append(new_pseudo)
+        
+    # Create a list, filters, to place Filter objects into.
+    filters = []
+    # Determine how many filters will be created.
+    total_filters = 1
+    for dimension in dimensions_pseudo:
+        total_filters *= len(dimension['parameters'])
+    # Generate a counter array to track which parameters to use.
+    counter = []
+    for i in range(len(dimensions_pseudo)):
+        counter.append(0)
+    # Structure a filter for each unique combination of parameters by dimension.
+    for i in range(total_filters):
+        new_filter = Filter(selections=[])
+        names = ""
+        abilities = ""
+        noabilities = ""
+        # Iterate the counter.
+        for j in range(len(dimensions_pseudo)):
+            counter[j] += 1
+            if counter[j] < len(dimensions_pseudo[j]['parameters']):
+                break
+            else:
+                counter[j] = 0        
+        # Add selected parameter combo data to helper strings.
+        for j in range(len(dimensions_pseudo)):
+            for include in dimensions_pseudo[j]['parameters'][counter[j]].include:
+                abilities += str(include) + "."
+            for exclude in dimensions_pseudo[j]['parameters'][counter[j]].exclude:
+                noabilities += str(exclude) + "."
+            names += dimensions_pseudo[j]['parameters'][counter[j]].name + "|"
+            new_selection = Selection(
+                dimension = dimensions_pseudo[j]['key'],
+                parameter = dimensions_pseudo[j]['parameters'][counter[j]].key
+                )
+            # Add Selection to the Filter.
+            new_filter.selections.append(new_selection)
+        # Add the entry to filters.
+        filter_string = query_base + "&filter="
+        if abilities != "":
+            filter_string += "abilities." + abilities[:-1] + "|"
+        if noabilities != "":
+            filter_string += "noabilities." + noabilities[:-1] + "|"
+        new_filter.name = names[:-1]
+        new_filter.string = filter_string[:-1]
+        filters.append(new_filter)
+        
+    # Determine how many pulls would be necessary for all ranks. This
+    # can be done in a resource-light manner by doing a dummy ranks pull
+    # with the filter 'abilities.9', as spell_ID 9 is not a spell and will
+    # filter out all actual ranks, returning only a total ranks value.
+    size_check_request = query_base + "&filter=abilities.9"
+    for attempt in range(10):
+        try:
+            total_ranks = pull_ranks(size_check_request)['total']
+            break
+        except PullFailedError:
+            pass
+    else:
+        raise ProcessFailureError(
+            "Could not get ranks count for Pull %d from user %s"
+            % (pull.key.id(), pull.parent().id()))
+    pull_counter = ((total_ranks / DEFAULT_LIMIT) + 1)
+    # Start a progress counter at 0.
+    progress_counter = 0
+    # For each filter:
+    for filter_ in filters:
+        filter_results = []
+        failed_pages = []
+        # Create a dict of dimension/parameter name tuples for the filter.
+        selections = {}
+        for selection in filter_.selections:
+            dimension = selection.dimension.get()
+            parameter = selection.parameter.get()
+            selections[dimension.name] = parameter.name
+        # For each pull required:
+        for page in range(1, (pull_counter + 1)):
+            # Add the page number to the filter query string.
+            query_string = filter_.string + "&page=" + str(page)
+            # Pull ranks and add to combined results.
+            try:
+                filter_results += pull_ranks(query_string)['rankings']
+            except PullFailedError as e:
+                failed_pages.append(page)
+                logging.error(e.msg)
+            # Increment the progress counter.
+            progress_counter += 1
+        # Add the tuples in selections to each rank dict
+        for rank in filter_results:
+            rank.update(selections)
+        # Add filter_results to results
+        results += filter_results
+        # Store failed pulls in NDB and add to failed_pulls
+        if len(failed_pages) > 0:
+            filter_.failed_pages = failed_pages
+            failed_pulls.append(filter_.put())
+            
+    # Finish Pull Work.
+    return finish_pull(pull, results, failed_pulls)
+        
+        
+# Pull ranks:
+def pull_ranks(query_string):
+    # Start with a timeout of 15 seconds.
+    timeout = 15
+    retry = True
+    # Make the ranks pull API request.
+    while retry == True:
+        try:
+            logging.info("Attempting %s" % query_string)
+            urlfetch.set_default_fetch_deadline(timeout)
+            response = urlfetch.fetch(query_string)
+            result = json.loads(response.content)
+            # Return the results.
+            return result
+        # If pull times out:
+        except urlfetch.Error:
+            logging.error("Pull failed.")
+            # Double the timeout.
+            timeout *= 2
+            # If this is the fourth attempt, give up.
+            if timeout > 60:
+                retry = False
+    # If it gets here it failed, so Raise PullFailedError.
+    raise PullFailedError("Pull failed with query string %s" % query_string)
+        
+        
+# Retry pull:
+def retry_pull(pull):
+    progress_counter = 0
+    pull_counter = 0
+    filters = ndb.get_multi(pull.failed_pulls)
+    failed_filters = filters
+    for filter_ in filters:
+        pull_counter += len(filter_.failed_pages)
+    succeeded_pulls = []
+    results = []
+    # Flag the pull as processing.
+    pull.status = 'Processing'
+    # For each failed filter:
+    for filter_ in filters:
+        # Create a dict of dimension/parameter name tuples for the filter.
+        selections = {}
+        for selection in filter_.selections:
+            dimension = selection.dimension.get()
+            parameter = selection.parameter.get()
+            selections[dimension.name] = parameter.name
+        # For each failed page:
+        for page in filter_.failed_pages:
+            # Add the page number to the filter query string.
+            query_string = filter_.string + "&page=" + str(page)
+            # Pull ranks and add to combined results.
+            try:
+                ranks += pull_ranks(query_string)
+                for rank in ranks:
+                    rank.update(selections)
+                filter_.failed_pages.remove(page)
+            except PullFailedError as e:
+                failed_pages.append(page)
+                logging.error(e.msg)
+            # Increment the progress counter.
+            progress_counter += 1
+        # If the filter was successfully completed, remove it from the list.
+        if len(filter_.failed_pages) == 0:
+            failed_filters.remove(filter_)
+            succeeded_pulls.append(filter_)
+    # Update any filters still in failed state.
+    if len(failed_filters) > 0:
+        failed_pulls = ndb.put_multi(failed_filters)
+    # Delete any successful filters.
+    if len(successful_filters) > 0:
+        ndb.delete_multi(succeeded_pulls)
+    # Finish Pull.
+    return finish_pull(pull, results, failed_pulls)
+        
+        
+# Finish Pull:
+def finish_pull(pull, results, failed_pulls):
+    pull_id = pull.key.id()
+    user_id = pull.key.parent().id()
+    filename_core = '/wclstats.appspot.com/' + str(pull_id) + "-" + str(user_id)
+    # Try to open existing results for this pull:
+    results_filename = filename_core + ".json"
+    try:
+        gcs_file = gcs.open(results_filename, 'r')
+        # Add the old results to the new results.
+        old_results = json.load(gcs_file)
+        results.append(old_results)
+        gcs_file.close()
+        # Overwrite the existing results file in cloud storage with the update.
+        gcs_file = gcs.open(results_filename, 'w', content_type='application/json')
+        gcs_file.write(json.dumps(results, ensure_ascii=False).encode('utf-8'))
+        gcs_file.close()
+    # If there are no results yet:
+    except gcs.NotFoundError:
+        # Create a results file in cloud storage.
+        gcs_file = gcs.open(results_filename, 'w', content_type='application/json')
+        gcs_file.write(json.dumps(results, ensure_ascii=False).encode('utf-8'))
+        gcs_file.close()
+    
+    # If failed_pulls is empty:
+    if len(failed_pulls) == 0:
+        csv_filename = filename_core + ".csv"
+        # Encode the results as a csv.
+        csv_results = csv_output(results, pull)
+        # Store the csv in cloud storage.
+        gcs_file = gcs.open(csv_filename, 'w', 
+                            content_type='text/csv; charset=UTF-8; header=present')
+        gcs_file.write(csv_results.encode('utf-8'))
+        gcs_file.close()
+        # Flag the pull as complete.
+        del pull.failed_pulls
+        pull.status = 'Ready'
+    # Else:
+    else:
+        # Store the failed_pulls list to the pull.
+        pull.failed_pulls = failed_pulls
+        # Flag the pull as incomplete.
+        pull.status = 'Incomplete'
+    
+    # Update the Pull in NDB.
+    pull.put()
     
     
 def csv_output(ranks, pull):
@@ -336,22 +330,18 @@ def csv_output(ranks, pull):
     request = main.Request.get_by_id(pull.request.id(), parent=pull.key.parent())
     dimensions = ndb.get_multi(request.dimensions)
     #If there's a trinket dimension, add it to the list of dimensions.
-    main.vislog('Adding trinket dimensions')
     if request.trinket_dimension != None:
         dimensions.insert(0, main.Dimension.get_by_id(request.trinket_dimension.id()))
     #Add all of the request dimensions add fields for the csv file.
-    main.vislog('Building field names')
     for dimension in dimensions:
         fieldnames.append(dimension.name)
     
     #Start the file with the field names.
-    main.vislog('Adding field names to file')
     for field in fieldnames:
         csvfile += field + ","
     csvfile = new_line(csvfile)
     
     #Take each rank and format it as csv.
-    main.vislog('Starting rank writes')
     for rank in ranks:
         for item in fieldnames:
             #If the item is the report link, take the fightID and turn it into
