+++ templates/elementmacros.html
@@ -183,6 +183,7 @@
 		<div class="row bottom-buffer">
 			<div class="col-sm-5">
 				<button class="btn btn-success" id="addTrinkets" type="button">Add Trinket Dimension</button>
+				<input hidden name="no_trinkets" value="1">
 				<hr>
 			</div>
 		</div>
@@ -348,7 +349,7 @@
 {%- endmacro %}
 
 <!-- mypulls.html macros -->
-{% macro pull_row_template(pull, request, difficulties, encounters, metrics) -%}
+{% macro pull_row_template(pull, request, difficulties, encounters, metrics, wcl_classes) -%}
 	<tr>
 		<td>
 			{% if pull.status == 'Ready' %}
@@ -358,12 +359,15 @@
 					</button>
 				</a>
 			{% elif pull.status == 'Processing' %}
-				<span class="label label-info">{{ pull.status }}</span>
+				<span class="btn btn-info">{{ pull.status }}</span>
 			{% elif pull.status == 'Queued' %}
-				<span class="label label-warning">{{ pull.status }}</span>
+				<span class="btn btn-warning">{{ pull.status }}</span>
+			{% elif pull.status == 'Incomplete' %}
+				<span class="btn btn-danger">{{ pull.status }}</span>
 			{% endif %}
 		</td>
-		<td>{{ request }}</td>
+		<td>{{ request.name }}</td>
+		<td>{{ wcl_classes[request.character_class - 1]['name'] }}/{{wcl_classes[request.character_class - 1]['specs'][pull.spec - 1]['name']}}
 		<td>{{ difficulties[pull.difficulty] }}</td>
 		<td>{{ encounters[pull.encounter] }}</td>
 		<td>{{ metrics[pull.metric] }}</td>
